separated sidebar and also changed the paths.
